+++ maptasker/src/userintr.py
@@ -261,15 +261,13 @@ class MyGui(customtkinter.CTk):
         self.sidebar_detail_option.configure(values=["0", "1", "2", "3", "4", "5"])
         self.sidebar_detail_option.set(str(DEFAULT_DISPLAY_DETAIL_LEVEL))
         self.display_detail_level = DEFAULT_DISPLAY_DETAIL_LEVEL
-        self.conditions = self.preferences = self.taskernet = self.debug = (
-            self.everything
-        ) = self.clear_settings = self.reset = self.restore = self.exit = self.bold = (
-            self.highlight
-        ) = self.italicize = self.underline = self.go_program = self.outline = (
-            self.rerun
-        ) = self.list_files = self.runtime = self.save = self.twisty = (
-            self.directory
-        ) = self.pretty = self.fetched_backup_from_android = False
+        self.conditions = self.preferences = self.taskernet = self.debug = self.everything = self.clear_settings = (
+            self.reset
+        ) = self.restore = self.exit = self.bold = self.highlight = self.italicize = self.underline = (
+            self.go_program
+        ) = self.outline = self.rerun = self.list_files = self.runtime = self.save = self.twisty = self.directory = (
+            self.pretty
+        ) = self.fetched_backup_from_android = False
         self.single_project_name = ""
         self.single_profile_name = ""
         self.single_task_name = ""
@@ -623,9 +621,7 @@ class MyGui(customtkinter.CTk):
         # We will prompt user for XML file if it hasn't already been loaded.
         name_entered = name_entered.strip()
         if name_entered and self.check_name(name_entered, my_name):
-            self.single_project_name = self.single_profile_name = (
-                self.single_task_name
-            ) = ""
+            self.single_project_name = self.single_profile_name = self.single_task_name = ""
 
             match my_name:
                 case "Project":
@@ -656,9 +652,6 @@ class MyGui(customtkinter.CTk):
 
         Returns:
             None
-
-        Note: This is a duplciate of the samd function in guiutils.py.
-        It can not be imported since it would cause a circular import.
         """
         self.textbox = customtkinter.CTkTextbox(self, height=650, width=250)
         self.textbox.grid(row=0, column=1, padx=(20, 0), pady=(20, 0), sticky="ew")
@@ -961,9 +954,7 @@ class MyGui(customtkinter.CTk):
             message_func = message_map.get(key)
             if message_func:
                 # Note: display_detail_level, file, font, indent, and single object name all return a message of 'None'.
-                message = (
-                    message_func()
-                )  # This calls the lambda function and takes a bit of time.
+                message = message_func()  # This calls the lambda function and takes a bit of time.
             # Catch bug where we have a key but no lambda function to process it.
             elif self.debug:
                 logger.debug(
@@ -1358,9 +1349,7 @@ class MyGui(customtkinter.CTk):
             # We have a file identified.  We now have to read it in.
             else:
                 filename_location = self.android_file.rfind(PrimeItems.slash) + 1
-                file_to_use = PrimeItems.program_arguments["android_file"][
-                    filename_location:
-                ]
+                file_to_use = PrimeItems.program_arguments["android_file"][filename_location:]
                 if not file_to_use:
                     file_to_use = self.android_file[filename_location:]
                 try:
@@ -1459,10 +1448,7 @@ class MyGui(customtkinter.CTk):
         window_position_attribute = f"{view_type}_window_position"
         window_title = f"{view_type.capitalize()} View"
 
-        if (
-            getattr(self, window_attribute) is None
-            or not getattr(self, window_attribute).winfo_exists()
-        ):
+        if getattr(self, window_attribute) is None or not getattr(self, window_attribute).winfo_exists():
             setattr(
                 self,
                 window_attribute,
@@ -1535,10 +1521,7 @@ class MyGui(customtkinter.CTk):
             None
         """
         # create window if its None or destroyed
-        if (
-            self.ai_analysis_window is None
-            or not self.ai_analysis_window.winfo_exists()
-        ):
+        if self.ai_analysis_window is None or not self.ai_analysis_window.winfo_exists():
             self.ai_analysis_window = TextWindow(
                 master=self,
                 window_position=self.ai_analysis_window_position,
@@ -1742,9 +1725,7 @@ class MyGui(customtkinter.CTk):
             except AttributeError:
                 temp_args[value] = ""
 
-        temp_args["ai_analyze"] = (
-            False  # Turn this off in event it was on from settings file.
-        )
+        temp_args["ai_analyze"] = False  # Turn this off in event it was on from settings file.
         _, _ = save_restore_args(temp_args, self.color_lookup, to_save=True)
 
         # force a reset of PrimeItems in mapit.py: initialize_everything.
@@ -1754,9 +1735,7 @@ class MyGui(customtkinter.CTk):
         # These flags are critical for the proper proceessing of the map.
         self.guiview = True  # Set it for save_settings
         PrimeItems.program_arguments["guiview"] = True  # Set it for mapit_all
-        PrimeItems.colors_to_use = (
-            self.color_lookup
-        )  # Make sure we have a color to use for mapit_all.
+        PrimeItems.colors_to_use = self.color_lookup  # Make sure we have a color to use for mapit_all.
 
         # Initialize a few things first
         if clear_names:
@@ -1815,9 +1794,7 @@ class EventHandlers:
         Returns:
             None
         """
-        self.parent = (
-            parent  # Save 'self' so widget event handlers can reference 'self'
-        )
+        self.parent = parent  # Save 'self' so widget event handlers can reference 'self'
 
     # Clear the message text box.
     def clear_messages_event(self) -> None:
@@ -1923,11 +1900,7 @@ class EventHandlers:
         )
 
         ### Port Number ###
-        android_port = (
-            "1821"
-            if the_view.android_port == "" or the_view.android_port is None
-            else the_view.android_port
-        )
+        android_port = "1821" if the_view.android_port == "" or the_view.android_port is None else the_view.android_port
         the_view.port_entry = the_view.port_label = None
         the_view.port_entry, the_view.port_label = the_view.display_label_and_input(
             "2-Port Number:",
@@ -2052,13 +2025,11 @@ class EventHandlers:
         if not ping_android_device(the_view, android_ipaddr, android_port):
             return
 
-        return_code, android_ipaddr, android_port, android_file = (
-            validate_or_filelist_xml(
-                the_view,
-                android_ipaddr,
-                android_port,
-                android_file,
-            )
+        return_code, android_ipaddr, android_port, android_file = validate_or_filelist_xml(
+            the_view,
+            android_ipaddr,
+            android_port,
+            android_file,
         )
 
         # Handle invalid file location or file not found.
@@ -2377,7 +2348,9 @@ class EventHandlers:
             - If the web browser is not supported, a message box is displayed.
             - If the web browser is supported, a message box is displayed with instructions for creating a new issue."""
         url = "//github.com/mctinker/Map-Tasker/issues"
-        issue_text = "Go to your browser and create a new issue or feature request, providing as much detail as possible."
+        issue_text = (
+            "Go to your browser and create a new issue or feature request, providing as much detail as possible."
+        )
         the_view = self.parent
         try:
             webbrowser.open(f"https:{PrimeItems.slash * 2}{url}", new=2)
@@ -2429,9 +2402,7 @@ class EventHandlers:
                 name_entered = "" if name_entered == "None" else name_entered
                 setattr(the_view, f"single_{my_name.lower()}_name", name_entered)
                 if name_entered:
-                    the_view.specific_name_msg = (
-                        f"Display only {my_name} '{name_entered}'."
-                    )
+                    the_view.specific_name_msg = f"Display only {my_name} '{name_entered}'."
                 else:
                     the_view.specific_name_msg = f"Display all {my_name}."
             else:
@@ -2541,10 +2512,7 @@ class EventHandlers:
         the_view.sidebar_detail_option.set(display_detail)
         the_view.inform_message("Display Detail Level", True, display_detail)
         # Disable twisty if detail level is less than 3
-        if (
-            the_view.twisty
-            and int(display_detail) < DISPLAY_DETAIL_LEVEL_all_parameters
-        ):
+        if the_view.twisty and int(display_detail) < DISPLAY_DETAIL_LEVEL_all_parameters:
             the_view.display_message_box(
                 f"Hiding Tasks with Twisty has no effect with Display Detail Level set to {display_detail}.  Twisty disabled!",
                 "Red",
@@ -2799,10 +2767,7 @@ class EventHandlers:
             the_view.twisty_checkbox,
             "Hide Task Details Under Twisty",
         )
-        if (
-            the_view.twisty
-            and int(the_view.display_detail_level) < DISPLAY_DETAIL_LEVEL_all_parameters
-        ):
+        if the_view.twisty and int(the_view.display_detail_level) < DISPLAY_DETAIL_LEVEL_all_parameters:
             the_view.display_message_box(
                 "This has no effect with Display Detail Level less than 3.  Display Detail Level set to 3!",
                 "Red",
@@ -3132,15 +3097,9 @@ class EventHandlers:
                 return
 
             # See if a valid API key was entered
-            if (
-                PrimeItems.ai[key] != value
-            ):  # If the key ent4ered doesn't matych what we already have.
+            if PrimeItems.ai[key] != value:  # If the key ent4ered doesn't matych what we already have.
                 # Validate the lngth of the key
-                if (
-                    value
-                    and key in apikeys_to_validate
-                    and not valid_api_key(key, value)
-                ):
+                if value and key in apikeys_to_validate and not valid_api_key(key, value):
                     error_msg = f"{key.replace('_key', '').title()} API key is invalid!"
                     my_gui.display_message_box(
                         error_msg,
@@ -3261,11 +3220,7 @@ class EventHandlers:
         if the_view.single_profile_name == "None or unnamed!":
             the_view.single_profile_name = ""
         # Do we have a single item identified?
-        if (
-            the_view.single_project_name
-            or the_view.single_profile_name
-            or the_view.single_task_name
-        ):
+        if the_view.single_project_name or the_view.single_profile_name or the_view.single_task_name:
             the_view.ai_analyze = True
             the_view.event_handlers.clear_messages_event()  # Clear out all displayed messages.
             the_view.display_message_box(
@@ -3387,9 +3342,7 @@ class EventHandlers:
         the_view.event_handlers.clear_messages_event()  # Clear out all displayed messages.
         # Go through loaded dictionary and display each line
         for key, value in changelog.items():
-            if (
-                "Older History" in value
-            ):  # Get out if we hit then of the the new version changes.
+            if "Older History" in value:  # Get out if we hit then of the the new version changes.
                 break
             if key == "version":
                 the_view.display_message_box(
@@ -3509,10 +3462,7 @@ class EventHandlers:
 
         # Check if we have a Project or Profile
         # If we don't already have Project, then get some XML.
-        if (
-            PrimeItems.tasker_root_elements["all_projects"]
-            or PrimeItems.tasker_root_elements["all_profiles"]
-        ):
+        if PrimeItems.tasker_root_elements["all_projects"] or PrimeItems.tasker_root_elements["all_profiles"]:
             # Process the diagram: builds the 'network' and then draws it in the GUI
             save_outline = guiview.outline
             guiview.outline = True
@@ -3544,12 +3494,8 @@ class EventHandlers:
             PrimeItems.program_arguments["doing_diagram"] = True  # Set it for mapit_all
 
             # Set our target objects since mapit-all will bypass setting these values
-            PrimeItems.program_arguments["single_project_name"] = (
-                guiview.single_project_name
-            )
-            PrimeItems.program_arguments["single_profile_name"] = (
-                guiview.single_profile_name
-            )
+            PrimeItems.program_arguments["single_project_name"] = guiview.single_project_name
+            PrimeItems.program_arguments["single_profile_name"] = guiview.single_profile_name
             PrimeItems.program_arguments["single_task_name"] = guiview.single_task_name
 
             # Re-invoke ourselves to force the html to be written
@@ -3566,12 +3512,12 @@ class EventHandlers:
                 return
 
             # Process the diagram file
-            diagram_dir = f"{os.getcwd()}{PrimeItems.slash}{DIAGRAM_FILE}"  # Get the directory from which we are running.
+            diagram_dir = (
+                f"{os.getcwd()}{PrimeItems.slash}{DIAGRAM_FILE}"  # Get the directory from which we are running.
+            )
             # Read the diagram file
             with open(str(diagram_dir), encoding="utf-8") as diagram_file:
-                diagram_data = [
-                    line.rstrip() for line in diagram_file
-                ]  # Read file into a list
+                diagram_data = [line.rstrip() for line in diagram_file]  # Read file into a list
 
                 # Display the diagram
                 guiview.diagramview = guiview.display_view("diagram", diagram_data)
@@ -3680,9 +3626,7 @@ class EventHandlers:
         )
 
         guiview.new_message_box(f"{title}\n\n{help_text}")
-        guiview.clear_messages = (
-            True  # Flag to tell display_message_box to clear the message box
-        )
+        guiview.clear_messages = True  # Flag to tell display_message_box to clear the message box
 
     def view_event(self: object, view_name: str) -> None:
         """
@@ -3780,9 +3724,7 @@ class EventHandlers:
                     textview.tag_items("found", f"Found: {search_input}   ")
 
                     # Add the list of matches to this entry in mygui.
-                    self.parent.items_for_selection["found"]["indecies"] = (
-                        textview.search_indecies
-                    )
+                    self.parent.items_for_selection["found"]["indecies"] = textview.search_indecies
 
                 # Mark located string by highlighting it.
                 textview.textview_textbox.tag_config(
@@ -3896,9 +3838,7 @@ class EventHandlers:
             textview.textview_textbox.see("1.0")
             display_msg = "Top"
             with contextlib.suppress(AttributeError, IndexError):
-                textview.search_current_line = textview.search_indecies[
-                    0
-                ]  # Point to first search hit.
+                textview.search_current_line = textview.search_indecies[0]  # Point to first search hit.
             textview.top = True
         else:
             # Go to bottom
@@ -3953,12 +3893,14 @@ class EventHandlers:
             bottom_line = connector["end_bottom"][0]
             for extra_bar in connector["extra_bars"]:
                 bottom_line = max(bottom_line, extra_bar[0])
-            seek_line = (
-                f"{bottom_line!s}.{connector['start_bottom'][1] - task_half_length!s}"
-            )
+            seek_line = f"{bottom_line!s}.{connector['start_bottom'][1] - task_half_length!s}"
             modifier = "bottom"
+
         # Display the Task
         textview.textview_textbox.see(seek_line)
+
+        # Display the appropriate jump-to-top or jump-to-bottom button.
+        CTkTextview.display_top_or_bottom_task_button(textview, connector)
 
         output_label(
             textview,